ml-backend/pyproject.toml
[project]
name = "peace-ml-backend"
version = "0.1.0"
description = "ML backend for PEACE endoscopy video analysis"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "opencv-python-headless>=4.9.0",
    "numpy>=1.26.0",
    "Pillow>=10.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
ml = [
    "torch>=2.0",
    "torchvision>=0.15",
]
dev = [
    "pytest>=8.0",
    "httpx>=0.27.0",
    "pytest-asyncio>=0.24.0",
]
